Allow mapping to partially disallowed objects
PPR mapping algorithm was refusing to map to objects
whose part was not within the allowed cpuset. This
typically occured when the cgroups were set up to
restrict processes to just half of all hyperthreads.
However, even in such case it makes perfect sense
to map (and bind) to a core/cache/numa etc.